Upload of Create Student Accounts
This is a GUI dialog to create student accounts. Follows the same
convention as CSVImport, but used for one-offs. Code in this was used
between the CSVImport file, and needs to be brought back. I will
probably end up with one unified file, that allows you to do wither bulk
CSV import, or individual, but for right now, these get the job done.