Upload of ResetPermissions
This is a quick (well not really) hack to reset all permissions on
folders on the fileserver. It is very dirty, but does work. It also
takes a long time on a lot of folders (2+ hours on ~250gig of teacher
data).
Basically run it as the domain Administrator account on the fileserver,
it takes ownership, makes sure inheritance is enabled, verifies the user
will have access, then sets ownership back to the user.
This relies on the folder name being the same as the user account in AD.
This fails if the user account was renamed, but the folder was not (e.x.
previous admin had changed some accounts of teacher who got marriade,
but did not change the homedirectory name. They were unable to get their
documents. I fixed this by renaming the folder, and updating the
homedirectory path in AD).